add tests and move vanishing route line to extension file
